refactor: move userBalances utilities to lib
