+++ scripts/userBalances/README.md
@@ -6,12 +6,14 @@ This script compares user balances between two different subgraphs to identify d
 
 - Fetches all users from both subgraphs in batches of 1000
 - Compares the following user data:
+
   - `gotchisOriginalOwned`
 
   - `portalsOwned` (unopened only)
   - `parcelsOwned`
   - `fakeGotchiCardBalances`
   - `fakeGotchiNFTTokens`
+
 - Identifies users that exist only in one subgraph
 - Detects differences in balances and values
 - Exports results to a JSON file
@@ -20,15 +22,16 @@ This script compares user balances between two different subgraphs to identify d
 
 ## Folder Structure
 
-The logic is split into several modules:
-- `compareUserBalances.ts` - main entry point.
-- `types.ts` - TypeScript interfaces.
-- `queries.ts` - GraphQL query strings.
-- `fetchers.ts` - data retrieval helpers.
-- `compare.ts` - comparison utilities.
-- `owners.ts` - functions for resolving original owners.
-- `utils.ts` - shared helpers.
-- `logger.ts` - simple colored logging helpers.
+The logic is split into several modules. The main script is
+`compareUserBalances.ts` and the supporting utilities live in the `lib` folder:
+
+- `lib/types.ts` - TypeScript interfaces.
+- `lib/queries.ts` - GraphQL query strings.
+- `lib/fetchers.ts` - data retrieval helpers.
+- `lib/compare.ts` - comparison utilities.
+- `lib/owners.ts` - functions for resolving original owners.
+- `lib/utils.ts` - shared helpers.
+- `lib/logger.ts` - simple colored logging helpers.
 
 ## Usage
