+++ scripts/userBalances/lib/compare.ts
